+++ tests/visual/test_visual_place.py
@@ -1,5 +1,3 @@
-import pikepdf
-
 import pdftl.api